Increased the size of the RX stream back to 256, otherwise it's not large enough to handle data coming in at high rate

+++ lrf_serial_comm.h
@@ -219,7 +219,7 @@ void set_diag_data_handler(LRFSerialCommApp *, void (*)(LRFDiag *, void *),
 void enable_shared_storage_dec_buf(LRFSerialCommApp *, bool);
 
 /** UART send function **/
-void uart_tx(LRFSerialCommApp *, uint8_t *, size_t);
+void uart_tx(LRFSerialCommApp *, uint8_t *, uint16_t);
 
 /** Send a command to the LRF **/
 void send_lrf_command(LRFSerialCommApp *, LRFCommand);
